flytt klasse til dedikert fil
